+++ developers/lib/components/customizer/css-handler.php
@@ -3,7 +3,7 @@
  * Adds the CSS from the Customizer options.
  *
  * @package     KnowTheCode\Developers
- * @since       1.0.0
+ * @since       1.1.0
  * @author      hellofromTonya
  * @link        https://knowthecode.io
  * @license     GNU General Public License 2.0+
@@ -12,24 +12,49 @@ namespace KnowTheCode\Developers\Customizer;
 
 add_action( 'wp_enqueue_scripts', __NAMESPACE__ . '\build_inline_css_from_customizer_settings' );
 /**
-* Checks the settings for the link color, and accent color.
-* If any of these value are set the appropriate CSS is output.
-*
-* @since 1.0.0
-*/
+ * Checks the settings for the link color, and accent color.
+ * If any of these value are set the appropriate CSS is output.
+ *
+ * @since 1.1.0
+ */
 function build_inline_css_from_customizer_settings() {
 	$prefix = get_settings_prefix();
 
-	$handle  = defined( 'CHILD_THEME_NAME' ) && CHILD_THEME_NAME ? sanitize_title_with_dashes( CHILD_THEME_NAME ) : 'child-theme';
+	$css = get_link_color_css( $prefix ) . get_accent_color_css( $prefix );
 
-	$color_link = get_theme_mod( $prefix . '_link_color', get_default_link_color() );
-	$color_accent = get_theme_mod( $prefix . '_accent_color', get_default_accent_color() );
+	if ( ! $css ) {
+		return;
+	}
 
-	$css = '';
+	$handle = defined( 'CHILD_THEME_NAME' ) && CHILD_THEME_NAME
+		? sanitize_title_with_dashes( CHILD_THEME_NAME )
+		: 'child-theme';
 
-	$css .= ( get_default_link_color() !== $color_link ) ? sprintf( '
+	wp_add_inline_style( $handle, $css );
 
-		a,
+}
+
+/**
+ * Get the Link Color Styles.
+ *
+ * @since 1.1.0
+ *
+ * @param string $prefix
+ *
+ * @return string
+ */
+function get_link_color_css( $prefix ) {
+	$color_link = get_theme_mod(
+		$prefix . '_link_color',
+		get_default_link_color()
+	);
+
+	if ( get_default_link_color() === $color_link ) {
+		return '';
+	}
+
+	return sprintf(
+		'a,
 		.entry-title a:focus,
 		.entry-title a:hover,
 		.genesis-nav-menu a:focus,
@@ -40,12 +65,32 @@ function build_inline_css_from_customizer_settings() {
 		.js nav button:focus,
 		.js .menu-toggle:focus {
 			color: %s;
-		}
-		', $color_link ) : '';
+		}',
+		$color_link
+	);
+}
 
-	$css .= ( get_default_accent_color() !== $color_accent ) ? sprintf( '
+/**
+ * Get the Accent Color styles.
+ *
+ * @since 1.1.0
+ *
+ * @param string $prefix
+ *
+ * @return string
+ */
+function get_accent_color_css( $prefix ) {
+	$color_accent = get_theme_mod(
+		$prefix . '_accent_color',
+		get_default_accent_color()
+	);
 
-		button:focus,
+	if ( get_default_accent_color() === $color_accent ) {
+		return '';
+	}
+
+	return sprintf(
+		'button:focus,
 		button:hover,
 		input:focus[type="button"],
 		input:focus[type="reset"],
@@ -61,11 +106,8 @@ function build_inline_css_from_customizer_settings() {
 		.sidebar .enews-widget input[type="submit"] {
 			background-color: %s;
 			color: %s;
-		}
-		', $color_accent, calculate_color_contrast( $color_accent ) ) : '';
-
-	if ( $css ) {
-		wp_add_inline_style( $handle, $css );
-	}
-
+		}',
+		$color_accent,
+		calculate_color_contrast( $color_accent )
+	);
 }