+++ developers/lib/components/woocommerce/views/theme-notice.php
@@ -1,0 +1,3 @@
+<div class="notice notice-info is-dismissible developer-woocommerce-notice">
+	<p><?php echo $notice_html; ?></p>
+</div>