+++ developers/lib/init.php
@@ -13,7 +13,7 @@ namespace KnowTheCode\Developers;
 /**
  * Initialize the theme's constants.
  *
- * @since 1.0.0
+ * @since 1.1.0
  *
  * @return void
  */
@@ -22,11 +22,12 @@ function init_constants() {
 	$child_theme = wp_get_theme();
 
 	define( 'CHILD_THEME_NAME', $child_theme->get( 'Name' ) );
-	define( 'CHILD_THEME_URL', $child_theme->get( 'ThemeURI' ) );
+	define( 'CHILD_THEME_AUTHOR_URL', $child_theme->get( 'ThemeURI' ) );
 	define( 'CHILD_THEME_VERSION', $child_theme->get( 'Version' ) );
 	define( 'CHILD_TEXT_DOMAIN', $child_theme->get( 'TextDomain' ) );
 
 	define( 'CHILD_THEME_DIR', get_stylesheet_directory() );
+	define( 'CHILD_THEME_URL', get_stylesheet_directory_uri() );
 	define( 'CHILD_CONFIG_DIR', CHILD_THEME_DIR . '/config/' );
 }